Create UserContent component
